✨ Bottom Up Modal - props 추가 & 변경
