+++ src/components/_common/BottomUpModal/index.tsx
@@ -6,15 +6,20 @@ import { ReactNode } from "react";
 
 interface Props {
   isModalOpen: boolean;
+  onClose: () => void;
+  navbarHeight: number;
   children: ReactNode;
 }
-export default function BottomUpModal({ isModalOpen, children }: Props) {
-  const [isOpen, setOpen] = useState(isModalOpen);
-
-  let h = (window.innerHeight - 47) / window.innerHeight;
+export default function BottomUpModal({
+  isModalOpen,
+  onClose,
+  navbarHeight,
+  children,
+}: Props) {
+  let h = (window.innerHeight - navbarHeight) / window.innerHeight;
 
   return (
-    <Sheet isOpen={isOpen} onClose={() => setOpen(false)} snapPoints={[h]}>
+    <Sheet isOpen={isModalOpen} onClose={onClose} snapPoints={[h]}>
       <Sheet.Container>
         <Sheet.Header />
         <Sheet.Content>{children}</Sheet.Content>
